requirements.txt
accelerate==0.33.0
diffusers
numpy==1.24.4
#torch==2.4.1
tqdm==4.66.2
transformers==4.40.1
#xformers==0.0.28.post1
einops==0.7.0
decord==0.6.0
sentencepiece==0.1.99
imageio
imageio-ffmpeg
ftfy
bs4
psutil # for live system stat monitoring in console
pytorch-msssim   # for managing interpolation files
